+++ packages/giu/src/inputs/checkbox.js
@@ -1,6 +1,7 @@
 // @flow
 
 import React from 'react';
+import PropTypes from 'prop-types';
 import { omit } from 'timm';
 import input, { INPUT_HOC_INVALID_HTML_PROPS } from '../hocs/input';
 
@@ -20,6 +21,7 @@ type PublicProps = {
   label?: any, // React components to be included in the checkbox's `label` element
   disabled?: boolean,
   styleLabel?: Object, // merged with the `label` style
+  skipTheme?: boolean,
   // all others are passed through to the `input` unchanged
 };
 // -- END_DOCS
@@ -29,6 +31,7 @@ const FILTERED_OUT_PROPS = [
   'label',
   'disabled',
   'styleLabel',
+  'skipTheme',
   ...INPUT_HOC_INVALID_HTML_PROPS,
 ];
 
@@ -47,6 +50,7 @@ class Checkbox extends React.Component {
   static defaultProps = {};
   props: Props;
   labelId: string;
+  refCheckbox: ?Object;
 
   constructor(props: Props) {
     super(props);
@@ -54,10 +58,19 @@ class Checkbox extends React.Component {
     cntId += 1;
   }
 
+  componentDidMount() {
+    if (this.context.theme === 'mdl' && this.refCheckbox) {
+      window.componentHandler.upgradeElement(this.refCheckbox);
+    }
+  }
+
   // ==========================================
   // Render
   // ==========================================
   render() {
+    if (!this.props.skipTheme && this.context.theme === 'mdl') {
+      return this.renderMdl();
+    }
     return this.props.label
       ? this.renderWithLabel()
       : this.renderInput('giu-checkbox');
@@ -95,7 +108,41 @@ class Checkbox extends React.Component {
       />
     );
   }
+
+  renderMdl() {
+    const inputProps = omit(this.props, FILTERED_OUT_PROPS);
+    return (
+      <label
+        ref={this.registerOuterRefMdl}
+        className="mdl-switch mdl-js-switch mdl-js-ripple-effect"
+        htmlFor={this.labelId}
+        style={style.wrapperMdl}
+      >
+        <input
+          ref={this.props.registerFocusableRef}
+          id={this.labelId}
+          className="mdl-switch__input"
+          type="checkbox"
+          checked={this.props.curValue}
+          {...inputProps}
+          disabled={this.props.disabled}
+        />
+        <span className="mdl-switch__label" style={style.labelMdl(this.props)}>
+          {this.props.label}
+        </span>
+      </label>
+    );
+  }
+
+  // ==========================================
+  registerOuterRefMdl = c => {
+    this.refCheckbox = c;
+    const { registerOuterRef } = this.props;
+    if (registerOuterRef) registerOuterRef(c);
+  };
 }
+
+Checkbox.contextTypes = { theme: PropTypes.any };
 
 // ==========================================
 // Styles
@@ -104,6 +151,15 @@ const style = {
   wrapper: {
     whiteSpace: 'nowrap',
   },
+  wrapperMdl: {
+    whiteSpace: 'nowrap',
+    width: 'initial',
+    marginRight: 10,
+  },
+  labelMdl: ({ styleLabel }) => ({
+    left: 16,
+    ...styleLabel,
+  }),
 };
 
 // ==========================================